+++ geotrek/settings/base.py
@@ -662,4 +662,9 @@ LOGGING = {
         },
     }
 }
+
+BLADE_CODE_TYPE = int
+BLADE_CODE_FORMAT = u"{signagecode}-{bladenumber}"
+LINE_CODE_FORMAT = u"{signagecode}-{bladenumber}-{linenumber}"
+
 SHOW_EXTREMITIES = False